Satisfy documentation copy governance
